requirements.txt
anyio==3.6.2
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.1
gcloud==0.17.0
googleapis-common-protos==1.59.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
jws==0.1.3
jwt==1.3.1
oauth2client==3.0.0
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.4.3
pydantic==1.10.7
pymongo==4.3.3
pyparsing==3.0.9
Pyrebase==3.0.27
python-decouple==3.8
python-dotenv==1.0.0
python-jwt==2.0.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.22.0
